% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfiles.R
\name{getLogs}
\alias{getLogs}
\title{\code{autoplot} a method that gives the recent log files}
\usage{
getLogs()
}
\value{
prints out a list of recent log files
}
\description{
\code{autoplot} a method that gives the recent log files
}
